requirements.txt
# Slack API
slack-sdk>=3.23.0

# Google Sheets API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Cloud Functions
functions-framework>=3.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
